Source Code for first functionality test, and JAR for new merged code
Source code for first successful test of merged code. I also included a new plugin (JAR file) to test having all code included, as it's own package. The hope being that it will make it so we only need one file and editing can be done more efficiently when testing.